# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "COMPATH API Documentation",
    "description": "COMPATH layihəsi üçün Auth, Tickets və Services API sənədləşdirilməsi",
    "version": "1.2.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "İstifadəçi girişi (Login)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@compath.az" },
                  "password": { "type": "string", "example": "123456" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Uğurlu giriş (Token qaytarılır)" },
          "400": { "description": "E-poçt və şifrə daxil edilməlidir" },
          "401": { "description": "Məlumatlar yanlışdır" }
        }
      }
    },
    "/api/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Yeni istifadəçi qeydiyyatı (Signup)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullname": { "type": "string", "example": "Təqdimat Test" },
                  "email": { "type": "string", "example": "user@compath.az" },
                  "password": { "type": "string", "example": "123456" }
                },
                "required": ["fullname", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "İstifadəçi uğurla yaradıldı" },
          "400": { "description": "Bütün sahələr doldurulmalıdır" },
          "409": { "description": "Bu e-poçt artıq qeydiyyatdan keçib" }
        }
      }
    },
    "/api/tickets": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Yeni şikayət yaratmaq (Create)",
        "security": [
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Kompüter problemi" },
                  "description": { "type": "string", "example": "Laboratoriyada kompüter açılmır" },
                  "department": { "type": "string", "example": "IT Dəstək" }
                },
                "required": ["title", "description", "department"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Şikayətiniz uğurla göndərildi" },
          "400": { "description": "Bütün sahələr doldurulmalıdır" },
          "401": { "description": "İcazə rədd edildi! Token yoxdur və ya etibarsızdır" }
        }
      },
      "get": {
        "tags": ["Tickets"],
        "summary": "İstifadəçinin öz şikayətlərinə baxması (Read)",
        "security": [
          { "BearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Şikayətlərin siyahısı uğurla gətirildi" },
          "401": { "description": "İcazə rədd edildi! Token yoxdur və ya etibarsızdır" }
        }
      }
    },
    "/api/tickets/all": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Bütün tələbələrin şikayətlərini gətir (Yalnız Admin)",
        "security": [
          { "BearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Bütün şikayətlərin siyahısı uğurla gətirildi" },
          "401": { "description": "Token yoxdur və ya etibarsızdır" },
          "403": { "description": "İcazə rədd edildi! Bu əməliyyat yalnız Adminlər üçündür" }
        }
      }
    },
    "/api/tickets/{ticket_id}": {
      "put": {
        "tags": ["Tickets"],
        "summary": "Şikayətin statusunu yeniləmək (Yalnız Admin)",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "Yenilənəcək şikayətin ID-si",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "example": "Həll olundu" }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status uğurla yeniləndi" },
          "400": { "description": "Status parametri boşdur və ya çatışmır" },
          "401": { "description": "Token yoxdur və ya etibarsızdır" },
          "403": { "description": "İcazə rədd edildi! Bu əməliyyat yalnız Adminlər üçündür" },
          "404": { "description": "Bu ID-yə uyğun şikayət tapılmadı" }
        }
      },
      "delete": {
        "tags": ["Tickets"],
        "summary": "Şikayəti silmək (Yalnız Admin)",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "Silinəcək şikayətin ID-si",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": { "description": "Şikayət uğurla silindi" },
          "401": { "description": "Token yoxdur və ya etibarsızdır" },
          "403": { "description": "İcazə rədd edildi! Bu əməliyyat yalnız Adminlər üçündür" },
          "404": { "description": "Bu ID-yə uyğun şikayət tapılmadı" }
        }
      }
    },
    "/api/services": {
      "get": {
        "tags": ["Services"],
        "summary": "Xidmətlərin siyahısını gətir (Hər kəs üçün)",
        "responses": {
          "200": { "description": "Xidmətlərin siyahısı" }
        }
      },
      "post": {
        "tags": ["Services"],
        "summary": "Yeni xidmət əlavə et (Yalnız Admin)",
        "security": [
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Akademik Arayış" },
                  "description": { "type": "string", "example": "Təhsil haqqında rəsmi arayışın verilməsi" }
                },
                "required": ["name", "description"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Xidmət uğurla yaradıldı" },
          "400": { "description": "Xidmət adı və təsviri daxil edilməlidir" },
          "401": { "description": "Token yoxdur və ya etibarsızdır" },
          "403": { "description": "İcazə rədd edildi! Bu əməliyyat yalnız Adminlər üçündür" }
        }
      }
    },
    "/api/services/{id}": {
      "delete": {
        "tags": ["Services"],
        "summary": "Xidməti sil (Yalnız Admin)",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Silinəcək xidmətin ID-si",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": { "description": "Xidmət uğurla silindi" },
          "401": { "description": "Token yoxdur və ya etibarsızdır" },
          "403": { "description": "İcazə rədd edildi! Bu əməliyyat yalnız Adminlər üçündür" },
          "404": { "description": "Xidmət tapılmadı" }
        }
      }
    }
  }
}